requirements.txt
pyzmq
protobuf
paho-mqtt
prompt-toolkit
rich
dronekit
psutil